renamed Int type to UInt64 - in real scenerio I would add u64 as a datatype or wrap it with a few functions to handle get() and set() but in this case would be a performance overhead.

+++ src/lib/core.ts
@@ -1,7 +1,7 @@
 import { EXT_NAME } from "../constants";
-import { type Int } from "../types";
+import { type UInt64 } from "../types";
 
-let ad_skip_counter: Int = 0;
+let ad_skip_counter: UInt64 = 0;
 
 export function skip_ad(player: HTMLVideoElement): void {
   if (isNaN(player.duration)) {
